Bug fixes in Summerize
Now I just take the minimum for each date, which is what I was trying to do previously.
Duplication in previous verion was due to missing `Inten (dB)` ==.
However this caused a new bug of missing entries due to not all files have the same min.

Found and fixed new bug due to not sorting by date prior to taking the top 3 entries.