New admin tools - matlab file copier and duplicate loader
Matlab finder - gets and copies the original matlab files; mainly a tool for Ben

Duplicate file loader - for days where the runs stop or there are 2 files for a rat, it automatically loads the 2nd file with no need to delete or anything. Side effect of erasing the 'old' data in rat_archive.